MIGRATIONS-1333: Documentation updates and multiple replay setup
