+++ deployment/cdk/opensearch-service-migration/lib/msk-utility-stack.ts
@@ -1,11 +1,11 @@
 import {StackPropsExt} from "./stack-composer";
-import {IVpc, Vpc} from "aws-cdk-lib/aws-ec2";
-import {CfnOutput, CfnWaitCondition, CfnWaitConditionHandle, CustomResource, Duration, Stack} from "aws-cdk-lib";
+import {IVpc} from "aws-cdk-lib/aws-ec2";
+import {CfnWaitCondition, CfnWaitConditionHandle, CustomResource, Duration, Stack} from "aws-cdk-lib";
 import {Construct} from "constructs";
 import {Effect, ManagedPolicy, PolicyDocument, PolicyStatement, Role, ServicePrincipal} from "aws-cdk-lib/aws-iam";
 import {NodejsFunction} from "aws-cdk-lib/aws-lambda-nodejs";
 import {Runtime} from "aws-cdk-lib/aws-lambda";
-import {AwsCustomResource, AwsCustomResourcePolicy, PhysicalResourceId, Provider} from "aws-cdk-lib/custom-resources";
+import {Provider} from "aws-cdk-lib/custom-resources";
 import * as path from "path";
 import {StringParameter} from "aws-cdk-lib/aws-ssm";
 
@@ -20,7 +20,7 @@ export class MSKUtilityStack extends Stack {
     constructor(scope: Construct, id: string, props: mskUtilityStackProps) {
         super(scope, id, props);
 
-        const mskARN = StringParameter.valueForStringParameter(this, `/migration/${props.stage}/mskClusterARN`)
+        const mskARN = StringParameter.valueForStringParameter(this, `/migration/${props.stage}/${props.defaultDeployId}/mskClusterARN`)
         let brokerEndpoints
         if (props.mskEnablePublicEndpoints) {
             const lambdaInvokeStatement = new PolicyStatement({
@@ -96,43 +96,51 @@ export class MSKUtilityStack extends Stack {
             brokerEndpoints = waitCondition.attrData.toString()
         }
         else {
-            // TODO reference ticket to replace this temporary workaround
-            try {
-                brokerEndpoints = StringParameter.valueFromLookup(this, `/migration/${props.stage}/mskBrokers`)
-            } catch (error) {
-                const mskGetBrokersCustomResource = getBrokersCustomResource(this, props.vpc, mskARN)
-                brokerEndpoints = mskGetBrokersCustomResource.getResponseField("BootstrapBrokerStringSaslIam")
-                //brokerEndpoints = mskGetBrokersCustomResource.getResponseField("BootstrapBrokerStringPublicSaslIam")
-            }
+            const mskUpdateConnectivityStatement = new PolicyStatement({
+                effect: Effect.ALLOW,
+                actions: ["kafka:GetBootstrapBrokers"],
+                resources: [mskARN]
+            })
+            const lambdaExecDocument = new PolicyDocument({
+                statements: [mskUpdateConnectivityStatement]
+            })
+
+            const lambdaExecRole = new Role(this, 'mskAccessLambda', {
+                assumedBy: new ServicePrincipal('lambda.amazonaws.com'),
+                description: 'Allow lambda to access MSK to retrieve brokers',
+                inlinePolicies: {
+                    mskAccessDoc: lambdaExecDocument,
+                },
+                // Required policies that the default Lambda exec role would add
+                managedPolicies: [
+                    ManagedPolicy.fromAwsManagedPolicyName("service-role/AWSLambdaVPCAccessExecutionRole"),
+                    ManagedPolicy.fromAwsManagedPolicyName("service-role/AWSLambdaBasicExecutionRole")
+                ]
+            });
+
+            const mskPublicLambda = new NodejsFunction(this, 'mskGetBrokersLambdaFunction', {
+                runtime: Runtime.NODEJS_18_X,
+                vpc: props.vpc,
+                handler: 'handler',
+                timeout: Duration.minutes(3),
+                entry: path.join(__dirname, 'lambda/msk-ordered-endpoints-handler.ts'),
+                role: lambdaExecRole,
+                environment: {MSK_ARN: mskARN}
+            });
+
+            const customResourceProvider = new Provider(this, 'customResourceProvider', {
+                onEventHandler: mskPublicLambda,
+            });
+
+            const customResource = new CustomResource(this, "mskGetBrokersCustomResource", {
+                serviceToken: customResourceProvider.serviceToken
+            })
+            brokerEndpoints = customResource.getAttString("BROKER_ENDPOINTS")
         }
         new StringParameter(this, 'SSMParameterMSKBrokers', {
             description: 'OpenSearch Migration Parameter for MSK Brokers',
-            parameterName: `/migration/${props.stage}/mskBrokers`,
+            parameterName: `/migration/${props.stage}/${props.defaultDeployId}/mskBrokers`,
             stringValue: brokerEndpoints
         });
     }
-}
-
-export function getBrokersCustomResource(scope: Construct, vpc: IVpc, clusterArn: string): AwsCustomResource {
-    const getBrokersCR = new AwsCustomResource(scope, 'migrationMSKGetBrokersCR', {
-        onCreate: {
-            service: 'Kafka',
-            action: 'getBootstrapBrokers',
-            parameters: {
-                ClusterArn: clusterArn,
-            },
-            physicalResourceId: PhysicalResourceId.of(Date.now().toString())
-        },
-        onUpdate: {
-            service: 'Kafka',
-            action: 'getBootstrapBrokers',
-            parameters: {
-                ClusterArn: clusterArn,
-            },
-            physicalResourceId: PhysicalResourceId.of(Date.now().toString())
-        },
-        policy: AwsCustomResourcePolicy.fromSdkCalls({resources: [clusterArn]}),
-        vpc: vpc
-    })
-    return getBrokersCR
 }
