+++ deployment/cdk/opensearch-service-migration/lib/service-stacks/kafka-zookeeper-stack.ts
@@ -21,7 +21,7 @@ export class KafkaZookeeperStack extends MigrationServiceCore {
     constructor(scope: Construct, id: string, props: KafkaZookeeperProps) {
         super(scope, id, props)
         let securityGroups = [
-            SecurityGroup.fromSecurityGroupId(this, "serviceConnectSG", StringParameter.valueForStringParameter(this, `/migration/${props.stage}/serviceConnectSecurityGroupId`)),
+            SecurityGroup.fromSecurityGroupId(this, "serviceConnectSG", StringParameter.valueForStringParameter(this, `/migration/${props.stage}/${props.defaultDeployId}/serviceConnectSecurityGroupId`)),
         ]
 
         const servicePort: PortMapping = {
