+++ deployment/cdk/opensearch-service-migration/test/stack-composer.test.ts
@@ -1,18 +1,14 @@
-import {App} from "aws-cdk-lib";
-import {StackComposer} from "../lib/stack-composer";
 import {Template} from "aws-cdk-lib/assertions";
-import {OpenSearchDomainStack} from "../lib/open-search-domain-stack";
+import {OpenSearchDomainStack} from "../lib/opensearch-domain-stack";
 import {createStackComposer} from "./test-utils";
 
 test('Test empty string provided for a parameter which has a default value, uses the default value', () => {
 
-    const app = new App({
-        context: {
-            domainName: ""
-        }
-    })
+    const contextOptions = {
+        domainName: ""
+    }
 
-    const openSearchStacks =  createStackComposer(app)
+    const openSearchStacks =  createStackComposer(contextOptions)
 
     const domainStack = openSearchStacks.stacks.filter((s) => s instanceof OpenSearchDomainStack)[0]
     const domainTemplate = Template.fromStack(domainStack)
@@ -22,27 +18,23 @@ test('Test empty string provided for a parameter which has a default value, uses
 
 test('Test invalid engine version format throws error', () => {
 
-    const app = new App({
-        context: {
-            // Should be OS_1.3
-            engineVersion: "OpenSearch_1.3"
-        }
-    })
+    const contextOptions = {
+        // Should be OS_1.3
+        engineVersion: "OpenSearch_1.3"
+    }
 
-    const createStackFunc = () => createStackComposer(app)
+    const createStackFunc = () => createStackComposer(contextOptions)
 
     expect(createStackFunc).toThrowError()
 })
 
 test('Test ES 7.10 engine version format is parsed', () => {
 
-    const app = new App({
-        context: {
-            engineVersion: "ES_7.10"
-        }
-    })
+    const contextOptions = {
+        engineVersion: "ES_7.10"
+    }
 
-    const openSearchStacks = createStackComposer(app)
+    const openSearchStacks = createStackComposer(contextOptions)
 
     const domainStack = openSearchStacks.stacks.filter((s) => s instanceof OpenSearchDomainStack)[0]
     const domainTemplate = Template.fromStack(domainStack)
@@ -51,13 +43,11 @@ test('Test ES 7.10 engine version format is parsed', () => {
 
 test('Test OS 1.3 engine version format is parsed', () => {
 
-    const app = new App({
-        context: {
-            engineVersion: "OS_1.3"
-        }
-    })
+    const contextOptions = {
+        engineVersion: "OS_1.3"
+    }
 
-    const openSearchStacks = createStackComposer(app)
+    const openSearchStacks = createStackComposer(contextOptions)
 
     const domainStack = openSearchStacks.stacks.filter((s) => s instanceof OpenSearchDomainStack)[0]
     const domainTemplate = Template.fromStack(domainStack)
@@ -66,29 +56,27 @@ test('Test OS 1.3 engine version format is parsed', () => {
 
 test('Test access policy is parsed for proper array format', () => {
 
-    const app = new App({
-        context: {
-            accessPolicies:
-                {
-                    "Version": "2012-10-17",
-                    "Statement": [
-                        {
+    const contextOptions = {
+        accessPolicies:
+            {
+                "Version": "2012-10-17",
+                "Statement": [
+                    {
                         "Effect": "Allow",
                         "Principal": {"AWS": "arn:aws:iam::12345678912:user/test-user"},
                         "Action": "es:ESHttp*",
                         "Resource": "arn:aws:es:us-east-1:12345678912:domain/test-os-domain/*"
-                        },
-                        {
-                            "Effect": "Allow",
-                            "Principal": {"AWS": "arn:aws:iam::12345678912:user/test-user2"},
-                            "Action": "es:ESHttp*",
-                            "Resource": "arn:aws:es:us-east-1:12345678912:domain/test-os-domain/*"
-                        }]
-                }
-        }
-    })
+                    },
+                    {
+                        "Effect": "Allow",
+                        "Principal": {"AWS": "arn:aws:iam::12345678912:user/test-user2"},
+                        "Action": "es:ESHttp*",
+                        "Resource": "arn:aws:es:us-east-1:12345678912:domain/test-os-domain/*"
+                    }]
+            }
+    }
 
-    const openSearchStacks = createStackComposer(app)
+    const openSearchStacks = createStackComposer(contextOptions)
 
     const domainStack = openSearchStacks.stacks.filter((s) => s instanceof OpenSearchDomainStack)[0]
     const domainTemplate = Template.fromStack(domainStack)
@@ -98,22 +86,20 @@ test('Test access policy is parsed for proper array format', () => {
 
 test('Test access policy is parsed for proper block format', () => {
 
-    const app = new App({
-        context: {
-            accessPolicies:
-                {
-                    "Version": "2012-10-17",
-                    "Statement": {
-                        "Effect": "Allow",
-                        "Principal": {"AWS": "*"},
-                        "Action": "es:ESHttp*",
-                        "Resource": "arn:aws:es:us-east-1:12345678912:domain/test-os-domain/*"
-                    }
+    const contextOptions = {
+        accessPolicies:
+            {
+                "Version": "2012-10-17",
+                "Statement": {
+                    "Effect": "Allow",
+                    "Principal": {"AWS": "*"},
+                    "Action": "es:ESHttp*",
+                    "Resource": "arn:aws:es:us-east-1:12345678912:domain/test-os-domain/*"
                 }
-        }
-    })
+            }
+    }
 
-    const openSearchStacks = createStackComposer(app)
+    const openSearchStacks = createStackComposer(contextOptions)
 
     const domainStack = openSearchStacks.stacks.filter((s) => s instanceof OpenSearchDomainStack)[0]
     const domainTemplate = Template.fromStack(domainStack)
@@ -123,92 +109,78 @@ test('Test access policy is parsed for proper block format', () => {
 
 test('Test access policy missing Statement throws error', () => {
 
-    const app = new App({
-        context: {
-            accessPolicies: {"Version":"2012-10-17"}
-        }
-    })
+    const contextOptions = {
+        accessPolicies: {"Version": "2012-10-17"}
+    }
 
-    const createStackFunc = () => createStackComposer(app)
+    const createStackFunc = () => createStackComposer(contextOptions)
 
     expect(createStackFunc).toThrowError()
 })
 
 test('Test access policy with empty Statement array throws error', () => {
 
-    const app = new App({
-        context: {
-            accessPolicies: {"Version":"2012-10-17", "Statement":[]}
-        }
-    })
+    const contextOptions = {
+        accessPolicies: {"Version": "2012-10-17", "Statement": []}
+    }
 
-    const createStackFunc = () => createStackComposer(app)
+    const createStackFunc = () => createStackComposer(contextOptions)
 
     expect(createStackFunc).toThrowError()
 })
 
 test('Test access policy with empty Statement block throws error', () => {
 
-    const app = new App({
-        context: {
-            accessPolicies: {"Version":"2012-10-17", "Statement":{}}
-        }
-    })
+    const contextOptions = {
+        accessPolicies: {"Version": "2012-10-17", "Statement": {}}
+    }
 
-    const createStackFunc = () => createStackComposer(app)
+    const createStackFunc = () => createStackComposer(contextOptions)
 
     expect(createStackFunc).toThrowError()
 })
 
 test('Test access policy with improper Statement throws error', () => {
 
-    const app = new App({
-        context: {
-            // Missing required fields in Statement
-            accessPolicies: {"Version":"2012-10-17", "Statement":[{"Effect": "Allow"}]}
-        }
-    })
+    const contextOptions = {
+        // Missing required fields in Statement
+        accessPolicies: {"Version": "2012-10-17", "Statement": [{"Effect": "Allow"}]}
+    }
 
-    const createStackFunc = () => createStackComposer(app)
+    const createStackFunc = () => createStackComposer(contextOptions)
 
     expect(createStackFunc).toThrowError()
 })
 
 test('Test invalid TLS security policy throws error', () => {
 
-    const app = new App({
-        context: {
-            tlsSecurityPolicy: "TLS_0_9"
-        }
-    })
+    const contextOptions = {
+        tlsSecurityPolicy: "TLS_0_9"
+    }
 
-    const createStackFunc = () => createStackComposer(app)
+    const createStackFunc = () => createStackComposer(contextOptions)
 
     expect(createStackFunc).toThrowError()
 })
 
 test('Test invalid EBS volume type throws error', () => {
 
-    const app = new App({
-        context: {
-            ebsVolumeType: "GP0",
-        }
-    })
+    const contextOptions = {
+        ebsVolumeType: "GP0",
+    }
 
-    const createStackFunc = () => createStackComposer(app)
+    const createStackFunc = () => createStackComposer(contextOptions)
 
     expect(createStackFunc).toThrowError()
 })
 
 test('Test invalid domain removal policy type throws error', () => {
 
-    const app = new App({
-        context: {
-            domainRemovalPolicy: "DELETE",
-        }
-    })
+    const contextOptions = {
+        domainRemovalPolicy: "DELETE",
+    }
 
-    const createStackFunc = () => createStackComposer(app)
+    const createStackFunc = () => createStackComposer(contextOptions)
 
     expect(createStackFunc).toThrowError()
 })
