+++ deployment/cdk/opensearch-service-migration/test/test-utils.ts
@@ -1,8 +1,16 @@
 import {Construct} from "constructs";
 import {StackComposer} from "../lib/stack-composer";
+import {App} from "aws-cdk-lib";
 
-export function createStackComposer(app: Construct) {
+export function createStackComposer(contextBlock: { [x: string]: (any); }) {
+    contextBlock.stage = "unit-test"
+    const app = new App({
+        context: {
+            contextId: "unit-test-config",
+            "unit-test-config": contextBlock
+        }
+    })
     return new StackComposer(app, {
-        env: {account: "test-account", region: "us-east-1"}, stage: "unittest"
+        env: {account: "test-account", region: "us-east-1"}
     })
 }
